Rewrite webserver, improve cog pass to fastapi, fix minor issues
