+++ .gitignore
@@ -12,3 +12,5 @@ test_bot.py
 code_test.py
 sample_reservation.json
 elo_vis.py
+Containerfile
+compose.yaml
